% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/authorize_box.R
\name{authorize_box}
\alias{authorize_box}
\title{Shortcut code to authorize Box.com API from API information}
\usage{
authorize_box()
}
\description{
Pulls API information from .Renviron without recall of specific code from the `boxr` package.
}
