% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparison_table.R
\name{comparison_table}
\alias{comparison_table}
\title{Generate comparison table of data from multiple time points}
\usage{
comparison_table(
  .data,
  comparison_variable,
  new_name = NULL,
  time_point = fy,
  cap_times = TRUE,
  data_format = "n_pct"
)
}
\arguments{
\item{.data}{The data frame.}

\item{comparison_variable}{Variable to compare response distribution across time points.}

\item{new_name}{Final name to display for the left (variable) column in the table. If not entered, will default to a sentence case version of the `comparison_variable`.}

\item{time_point}{Name of time point variable in dataset. Defaults to "fy" (without quotes).}

\item{cap_times}{Do you want to force the time point columns to be renamed as all caps (T/F)? Defaults to TRUE.}

\item{data_format}{Format of data presented in table. Can be one of "n" (only display n), "pct" (only display percent), or "n_pct" (display both n and percent; default).}
}
\value{
A data frame.
}
\description{
Produces a table with response options (left) and time points (top) in n (%) format.
}
