% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crosstab.R
\name{crosstab}
\alias{crosstab}
\title{Generate a crosstab between an outcome and a demographic variable}
\usage{
crosstab(.data, question, demographic)
}
\arguments{
\item{.data}{The data frame.}

\item{question}{Outcome variable.}

\item{demographic}{Demographic variable to stratify data from outcome variable.}
}
\value{
A tabyl.
}
\description{
Produces a tabyl with demographic responses on the left and the various answer options to the outcome variable on the top. Also includes a column to the right with the n for each demographic response item.
}
