% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctlr.R
\docType{package}
\name{ctlr}
\alias{ctlr-package}
\alias{ctlr}
\title{Streamlined Cleaning, Analysis, and Visualization Tools}
\description{
I'll ensure a longer description goes here eventually.
}
\author{
Jay Cutler \email{jacobmcutler@outlook.com}
}
