% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frequency_table.R
\name{frequency_table}
\alias{frequency_table}
\title{Generate a basic table with n and percent.}
\usage{
frequency_table(.data, frequency_variable, ordering)
}
\arguments{
\item{.data}{The data frame.}

\item{frequency_variable}{Variable of interest.}

\item{ordering}{Choose to order data from greatest to lowest percent ("pct") or by factor order ("fct"). Defaults to "pct". (Note: if there are no factor levels set and "fct" is selected, data will arrange variable of choice in alphabetical order.)}
}
\value{
A tabyl.
}
\description{
Produces a tabyl with n and percent for a given variable. Excludes NAs and by default will order data from highest to lowest percent
}
