% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handle_fonts.R
\name{handle_fonts}
\alias{handle_fonts}
\title{Register family of fonts with systemfonts()}
\usage{
handle_fonts(font_name)
}
\arguments{
\item{font_name}{Name of the font as it would appear in a word processing software (e.g., "Helvetica Neue", "Open Sans").}
}
\description{
Loads all font styles within a font family, already installed on your system, into R to be used within {ggplot2}.
}
