% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mff_table.R
\name{mff_table}
\alias{mff_table}
\title{MFF Table}
\usage{
mff_table(.data)
}
\arguments{
\item{.data}{The data frame.}
}
\value{
A flextable object.
}
\description{
Display data frame in MFF flextable format.
}
