% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pct_over_100.R
\name{pct_over_100}
\alias{pct_over_100}
\title{Calculate n and percent for checklist-style question data}
\usage{
pct_over_100(.data, .cols, id = "id")
}
\arguments{
\item{.data}{The data frame.}

\item{.cols}{Columns to select with data.}

\item{id}{Name of identifier column. Defaults to "id".}
}
\value{
A data frame.
}
\description{
Produces a table with your variable of interest, n, and percentages where multiple columns are pivoted longer and there may be more than one option (column) per row/respondent ID.
}
