% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pretty_header.R
\name{pretty_header}
\alias{pretty_header}
\title{Display a heading "bar" as a formatted table for Word documents}
\usage{
pretty_header(
  header_text,
  bg_color = "#13334c",
  header_font = "Montserrat",
  text_color = "white"
)
}
\arguments{
\item{header_text}{The header name, as a string, that you would like displayed in the table.}

\item{bg_color}{The background color of the table. Defaults to "#13334c" (dark navy).}

\item{header_font}{The font of the header. Defaults to "Montserrat".}

\item{text_color}{The color of the header text. Defaults to white.}
}
\value{
A flextable object.
}
\description{
Generates a heading as a table for Word documents using `flextable`. The colors and fonts utilized are customizable and any additional modifications can be added as additional `flextable` functions to override any defaults set by this function.
}
