% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_fill_mff.R
\name{scale_fill_mff}
\alias{scale_fill_mff}
\title{Use an MFF palette for a color argument.}
\usage{
scale_fill_mff(
  palette = "Blue",
  discrete = TRUE,
  reverse = FALSE,
  contrast = "high"
)
}
\arguments{
\item{palette}{Selected palette. All palettes are 9 colors except for "Brand" (3) and "DivergingEven" (8). Options include: Brand, Blue, Teal, Orange, Green, Gradient, Diverging, and DivergingEven.}

\item{discrete}{Is this for a set of discrete or continuous values (T/F)? Defaults to TRUE.}

\item{reverse}{Do you want the palette to display in reverse order (T/F)? Defaults to FALSE.}

\item{contrast}{Level of contrast desired between colors in plot. Options are "high", "med", and "low". Defaults to "high".}
}
\description{
Use an MFF palette for a color argument.
}
