% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_qualtrics_api.R
\name{set_qualtrics_api}
\alias{set_qualtrics_api}
\title{Choose from loaded MFF Qualtrics APIs}
\usage{
set_qualtrics_api(account)
}
\arguments{
\item{account}{Name of MFF Qualtrics account to use. Can be one of "snap-ed", "ntae", or "other". Must ensure that API keys are named in lowercase and underscores with these names, and base URLs are named either with "mff" or "ntae" and underscores in the same format (e.g., "ntae_qualtrics_api_key", "mff_qualtrics_base_url").}
}
\description{
Loads specified API information to pull down data from Qualtrics.
}
