% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctlr.R
\docType{package}
\name{ctlr}
\alias{ctlr-package}
\alias{ctlr}
\title{ctlr: Shortcuts for data cleaning, analysis, and visualization}
\description{
This package utilizes functions from the {tidyverse} as well as {flextable}, {qualtRics}, and {boxr} to offer function "shortcuts" for simple data cleaning, wrangling similar types of survey data, visualizing data in tabular and graphical forms, and connecting and utilizing APIs.
}
\author{
Jay Cutler
\email{jacobmcutler@outlook.com}
}
